additional functionality and fixes
