+++ src/presenter/new-point-button-presenter.js
@@ -1,0 +1,15 @@
+import {render} from '../framework/render';
+
+export default class NewPointButtonPresenter {
+  #container = null;
+  #component = null;
+
+  constructor({container, component}) {
+    this.#container = container;
+    this.#component = component;
+  }
+
+  init() {
+    render(this.#component, this.#container);
+  }
+}
