+++ src/template/loading-template.js
@@ -1,0 +1,3 @@
+export default function createLoadingTemplate() {
+  return '<p class="trip-events__msg">Loading...</p>';
+}
