+++ src/template/new-point-button-template.js
@@ -1,0 +1,3 @@
+export default function createNewPointButtonTemplate() {
+  return '<button class="trip-main__event-add-btn btn btn--big btn--yellow" type="button">New event</button>';
+}
