+++ src/template/trip-info-template.js
@@ -1,0 +1,5 @@
+import {NO_POINTS_TEXT_TYPE} from '../const';
+
+export default function createTripInfoTemplate(filterType) {
+  return `<p class="trip-events__msg">${NO_POINTS_TEXT_TYPE[filterType]}</p>`;
+}
