+++ src/view/filter-view.js
@@ -1,30 +1,5 @@
 import AbstractView from '../framework/view/abstract-view';
-import {camelizer} from '../utils';
-
-function createFilterTemplate(filters, currentFilterType) {
-  return (
-    `
-    <div class="trip-main__trip-controls  trip-controls">
-      <div class="trip-controls__filters">
-        <h2 class="visually-hidden">Filter events</h2>
-        <form class="trip-filters" action="#" method="get">
-
-          ${filters.map(({type, isDisabled}) => `
-            <div class="trip-filters__filter">
-              <input id="filter-${type}" class="trip-filters__filter-input  visually-hidden" ${isDisabled ? 'disabled' : ''} type="radio" name="trip-filter" value="${type}" ${type === currentFilterType ? 'checked' : ''}>
-              <label class="trip-filters__filter-label" for="filter-${type}">
-                ${camelizer(type)}
-              </label>
-            </div>
-          `).join('')}
-
-            <button class="visually-hidden" type="submit">Accept filter</button>
-        </form>
-        </div>
-    </div>
-    `
-  );
-}
+import createFilterTemplate from '../template/filter-template';
 
 export default class FilterView extends AbstractView{
   #filters = null;
