+++ src/view/info-view.js
@@ -1,23 +1,5 @@
 import AbstractView from '../framework/view/abstract-view';
-
-function createInfoTemplate(info) {
-  const { destinationsString, datesString, total } = info;
-  return (
-    `
-      <section class="trip-main__trip-info  trip-info">
-        <div class="trip-info__main">
-          <h1 class="trip-info__title">${destinationsString}</h1>
-
-          <p class="trip-info__dates">${datesString}</p>
-        </div>
-
-        <p class="trip-info__cost">
-          Total: &euro;&nbsp;<span class="trip-info__cost-value">${total}</span>
-        </p>
-      </section>
-    `
-  );
-}
+import createInfoTemplate from '../template/info-template';
 
 export default class InfoView extends AbstractView {
   #info = null;
