+++ src/view/loading-view.js
@@ -1,8 +1,5 @@
 import AbstractView from '../framework/view/abstract-view.js';
-
-function createLoadingTemplate() {
-  return '<p class="trip-events__msg">Loading...</p>';
-}
+import createLoadingTemplate from '../template/loading-template';
 
 export default class LoadingView extends AbstractView {
   get template() {
