+++ src/view/new-point-button-view.js
@@ -1,8 +1,5 @@
 import AbstractView from '../framework/view/abstract-view.js';
-
-function createNewPointButtonTemplate() {
-  return '<button class="trip-main__event-add-btn btn btn--big btn--yellow" type="button">New event</button>';
-}
+import createNewPointButtonTemplate from '../template/new-point-button-template';
 
 export default class NewPointButtonView extends AbstractView {
   #handleClick = null;
