--- src/view/point-list-view.js
@@ -1,11 +1,0 @@
-import AbstractView from '../framework/view/abstract-view.js';
-
-function createNewTaskButtonTemplate() {
-  return '<ul class="trip-events__list"></ul>';
-}
-
-export default class PointListView extends AbstractView{
-  get template() {
-    return createNewTaskButtonTemplate();
-  }
-}
