+++ src/view/sort-view.js
@@ -1,43 +1,5 @@
 import AbstractView from '../framework/view/abstract-view';
-
-function createSortTemplate({sortType}) {
-  return (
-    `
-      <form class="trip-events__trip-sort  trip-sort" action="#" method="get">
-        <div class="trip-sort__item  trip-sort__item--day">
-          <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day"
-            ${sortType === 'day' ? 'checked' : ''}
-          >
-          <label class="trip-sort__btn" for="sort-day" data-sort-type="day">Day</label>
-        </div>
-
-        <div class="trip-sort__item  trip-sort__item--event">
-          <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event" disabled>
-          <label class="trip-sort__btn" for="sort-event">Event</label>
-        </div>
-
-        <div class="trip-sort__item  trip-sort__item--time">
-          <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-time"
-            ${sortType === 'time' ? 'checked' : ''}
-          >
-          <label class="trip-sort__btn" for="sort-time" data-sort-type="time">Time</label>
-        </div>
-
-        <div class="trip-sort__item  trip-sort__item--price">
-          <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price"
-            ${sortType === 'price' ? 'checked' : ''}
-          >
-          <label class="trip-sort__btn" for="sort-price" data-sort-type="price">Price</label>
-        </div>
-
-        <div class="trip-sort__item  trip-sort__item--offer">
-          <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-offer" disabled>
-          <label class="trip-sort__btn" for="sort-offer">Offers</label>
-        </div>
-      </form>
-    `
-  );
-}
+import createSortTemplate from '../template/sort-template';
 
 export default class SortView extends AbstractView {
   #onSortTypeChange = null;
